Revert to add all child tasks w. estimated time
Only condition for a task to be added from selection to Time2 is that the task have an estimated time set. That's it.

At first I thought that "add only selection" was the simplest. And in practice I probably will have all tasks selected every time I run the script, but this is even simpler; I merely have to select all relevant parent tasks in project and not worry about all subtasks being expanded in the view.